Add `tlsSecret` to `Host`
Must be added alongside `tlsContext`. For compatibility with Emissary
Ingress 2.x. See https://www.getambassador.io/docs/emissary/latest/topics/running/host-crd/#tlscontext-links-to-a-tlscontext-for-additional-configuration

```
tlsContext specifies a TLSContext to use for additional TLS information. Note that you must still define tlsSecret for TLS termination to happen. It is an error to supply both tlsContext and tls.
```